requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
datasets>=2.14.0
wandb>=0.15.0
pyyaml>=6.0
tqdm>=4.65.0
scipy>=1.10.0
accelerate>=0.21.0
bitsandbytes>=0.41.0